Commit latest snapshot.
I've determined why the process blocks after all tests completes. It's
because you have to disconnect from mongodb, or the event loop with be
blocked. This makes sense, but what doesn't is why jasmine doesn't
provide hooks for beforeAll and afterAll. It's here where you'd want to
connect and disconnect from mongodb. Next step is to install the plugin
that monkey patches jasmine to provide this functionality.

* https://github.com/mhevery/jasmine-node/issues/241
* https://github.com/nonplus/jasmine-beforeAll
* https://github.com/pivotal/jasmine/pull/56